created Auth.js component, created page not found route and component
