+++ src/components/NotFound.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class NotFound extends React.Component {
+  render() {
+    return (
+      <div>
+        <h3>Oops! Page not found...</h3>
+      </div>
+    );
+  }
+}
+
+export default NotFound;
