+++ src/components/Router.js
@@ -3,6 +3,7 @@ import { Switch, Route, BrowserRouter as Router } from 'react-router-dom';
 import Home from './Home';
 import Login from './Login';
 import Dashboard from './Dashboard';
+import NotFound from './NotFound';
 
 const AppRouter = () => {
   return (
@@ -11,6 +12,7 @@ const AppRouter = () => {
         <Route exact path="/" component={Home} />
         <Route path="/login" component={Login} />
         <Route path="/dashboard" component={Dashboard} />
+        <Route path="/notfound" component={NotFound} />
       </Switch>
     </Router>
   );
